requirements.txt
python-telegram-bot[job-queue]
ccxt
pandas
ta
websockets